v.0.5
Features:
- create workorder with CSVs and processed
- workorder filters
- save and retrieve CSV files stored
- delete orders in workorder (delete in cascade)
- staff model
- new workflow to introduce the work orders
TODO:
- save and retrieve invoice files (PDFs)
Known bugs:
- WorkOrder model: save method: It doesn’t get the correct acronym

+++ staff/workOrders/admin.py
@@ -1,26 +1,16 @@
 from django.contrib import admin
-from workOrders.models import WorkOrder, Order
-
-import pandas as pd
-import string
-from datetime import datetime
-
-def getAcronym(filename):
-
-    return filename[27:33]
-
-def removeDummyProyects(report):
-
-    return report[report.Project.str.contains('^X')==False]
+from workOrders.models import WorkOrder, Order, Staff
 
 
-def getReport(fileName):
-    report = pd.read_csv(fileName, sep=';', names=['Date','Hours','Project','Phase','SubPhase','Task','OnCall','CallIn'])
-    report['staffId']=getAcronym(fileName.name)
-    report['Hours'] = report['Hours'].apply(lambda x: float(string.replace(x,',','.')))
-    report['Date'] = [datetime.strptime(x, '%d/%m/%Y') for x in report['Date']]
-    report['Period'] = [str(_date.year)+'-'+str(_date.month)+'-01' for _date in report['Date']]
-    return removeDummyProyects(report)
+class StaffAdmin(admin.ModelAdmin):
+    fieldsets = [
+        (None, {'fields':['staffId','name','surname','company', 'start_Date', 'end_Date', 'daily_Rate']})
+    ]
+    list_display = ('staffId', 'name','surname',  'company', 'start_Date', 'end_Date', 'daily_Rate')
+
+    list_filter = ['company', 'end_Date', 'staffId']
+
+
 
 class OrderInline(admin.TabularInline):
     model = Order
@@ -38,21 +28,27 @@ class WorkOrderAdmin(admin.ModelAdmin):
         """ 
         """
         for workorder in queryset:
+            if (workorder.total_No_Taxes == None) :
+                break
             workorder.deleteAllOrders()
-            workOrderProcess = getReport(workorder.woFile).groupby(['staffId','Project','Period'], as_index=False).sum()
+            workOrderProcess = workorder.getReport().groupby(['staffId','Project','Period'], as_index=False).sum()
             totHours = workOrderProcess['Hours'].sum()
             bill = float(workorder.total_No_Taxes)
             workOrderProcess['Percentage'] = workOrderProcess['Hours']/totHours
             workOrderProcess['Amount'] = workOrderProcess['Percentage']*bill
+
             for orderProcess in workOrderProcess.iterrows():
                 order = Order(workOrder=workorder, staffId=orderProcess[1]['staffId'], 
                                 project = orderProcess[1]['Project'],
                                 woDate=orderProcess[1]['Period'], hours = orderProcess[1]['Hours'], 
                                 perctHours = orderProcess[1]['Percentage'], total = orderProcess[1]['Amount'])
                 order.save()
+            
             workorder.state = WorkOrder.PROCESSED
-            workorder.staffId = getAcronym(workorder.woFile.name)
-            workorder.save(force_update=True, update_fields=['state','staffId'])
+            #workorder.staffId = Staff.object.get(getAcronym(workorder.woFile.name))
+            workorder.save(force_update=True)
+            #workorder.save(force_update=True, update_fields=['state','staffId','staffId_id'])
+            print "WorkOrder PROCESSED"
 
     process.short_description = "Process selected workorders"
 
@@ -63,13 +59,15 @@ class WorkOrderAdmin(admin.ModelAdmin):
     delete.short_description = 'Delete selected workorders'
 
 
-    readonly_fields = ['staffId','woDate']
+    readonly_fields = ['staffId','woDate','getStaffName']
 
     fieldsets = [
-        (None, {'fields': ['staffId', 'state', 'woDate', 'total_No_Taxes', 'woFile']}),
+        ('Staff', {'fields': ['staffId','getStaffName']}),
+        (None, {'fields': ['state', 'woDate', 'total_No_Taxes', 'woFile']}),
+        
     ]
 
-    list_display = ('staffId', 'state','woDate', 'total_No_Taxes', 'woFile')
+    list_display = ('getStaffName','staffId', 'state','woDate', 'total_No_Taxes', 'woFile')
 
     inlines = [OrderInline]
 
@@ -83,3 +81,4 @@ admin.site.disable_action('delete_selected')
 # Register your models here.
 admin.site.register(WorkOrder, WorkOrderAdmin)
 admin.site.register(Order, OrderInline)
+admin.site.register(Staff, StaffAdmin)
